+++ src/clj_rest_web_app/core.clj
@@ -1,6 +1,18 @@
 (ns clj-rest-web-app.core
+  (:require [compojure.core :refer :all]
+            [ring.adapter.jetty :as ring]
+            [ring.middleware.defaults :refer [wrap-defaults site-defaults]])
   (:gen-class))
 
-(defn -main
-  [& args]
-  (println "Hello World"))
+(defroutes app-routes
+  (GET "/" [] "<h1>Hello World</h1>"))
+
+(def application (wrap-defaults app-routes site-defaults))
+
+(defn start [port]
+  (ring/run-jetty application {:port port
+                               :join? false}))
+
+(defn -main []
+  (let [port (Integer. (or (System/getenv "PORT") "8080"))]
+    (start port)))