+++ CHANGELOG.md
@@ -5,5 +5,6 @@ All notable changes to this project will be documented in this file. This change
 ### Added
 - Add support for Docker
 - Add database migration tooling
+- HTTPify the application
 
 [Unreleased]: https://github.com/bakku/clj-rest-web-app/compare/5a1bf27...HEAD